Apply toggle setting on page load
If the toggle is checked when selecting a taxa or refreshing
the page then the state is not respected and you have to untick
the checkbox and then recheck it. This code applies it when the
tinytax code is attached.

Note that the code has been implemented in a way that ensures we
only alter the elements within the context that have the class
toggleable. This follows the rest of the code which makes sure it
uses the vid value throughout to only alter the tree it is attached
to. This seems like a clear choice though it is not doc'd, however,
it goes against the cookie value which is a single cookie for all
vids. In the absense of comments explaining this, the code has been
written so that it only acts on the tree it is attached to and not
all .toggleable elements.