Patch httprl to accept empty chunks
In httprl there is a condition that stops the process from reading
the stream of data coming out of an fp if an empty chunk is read,
regardless of whether the eof has been reached. This is a bit odd
and is causing at least one noticed issue on scratchpads, namely
that BHL API calls we make don't always work.

This issue has been raised against the httprl drupal project here:
https://www.drupal.org/project/httprl/issues/2818613 and the patch
uploaded is what is included in this commit (I applied the changes
manually and then made a new patch to ensure all the paths were
correct). This is the original patch:
https://www.drupal.org/files/issues/httprl-empty_chunk_continue-2818613-2-D7.patch.
The patch has been in the "needs review" state since 2016 and therefore
it doesn't look likely it will get merged (the repo doesn't really
seem to be actively maintained anymore).

By default the new flag that enables this empty chunk skipping is
turned off to avoid altering any existing uses in the codebase,
however, it can be turned on if needed by passing
'ignore_empty_chunk' => true as an option.