Use the new ignore_empty_chunk option in BHL API calls
For unknown reasons making two calls to the BHL API in a row
causes either the BHL API or httprl to generate an empty chunk
when reading the data the second time through. Using this
flag avoids this empty chunk being a problem by ignoring empty
chunks and carrying on reading until timeout/eof rather than
breaking out when reading an empty chunk.