Ignore empty strings in URL column on import
If the url is an empty string then the MediaFeedsInternetProvider
will throw a MediaInternetNoHandlerException because it can't find
a handler for it. Presumably this is desired under normal circumstances,
but during an import the cells values are empty strings so this error
is thrown all the time. Therefore, preempt it before it happens!