Allow the taxonomy overview admin view to show additional operations
The operations that can be shown are hard coded as 'add', 'list' and 'edit' but
this means we can't add any more. This patch adds an additional key to each
vocabulary child of the form to list the operations that should be shown with it.
By default these are 'add', 'list' and 'edit'.

To add a new operation to a vocabulary, simply add a new entry into the #opkeys
array in the vocabulary form child and then add a corresponding key with the
operation array configuration as well. For example:

```
$child['whatever'] = array(
  '#type' => 'link',
  '#title' => t('Does something!'),
  '#href' => 'somewhere/on/the/site'
);
$child['#opkeys'][] = 'whatever';
```