Use unshift instead of push when adding the fids to the form values array
I'm not 100% clear why this works, but when using push the values do not
come through in the php. My guess would be that adding the values to the
end of the form values array places them after some value that the php
code recognises as the end of the form values and the start of other stuff
that it can ignore. For example, the form values I've been working with on
a biblio node has over a 1000 elements, the first few are the form values
and then the other like 990 are "ajax_html_ids[]" and "ajax_page_state"
items. Any way, adding the values to the start of list means they come
through in the php and it all works.