Remove duplicates by using distinct query setting
Files associated with a taxonomic term seem to be duplicated because multiple
relationship rows are created in the field_data_field_taxonomic_name table
to reflect the various languages available on the scratchpad. This seems a bit
odd but presumably is the desired behaviour and I managed to find other users
with the same kind of problem where duplicated are shown based on a language
issue. One suggested fix is to use the field language setting on the view but
we already have this set (Field Language:Current user's language) so instead
I've added a distinct to the query to avoid duplicate file ids form being
returned and then displayed.