Allow multiple import columns to map to a single media field
This shows up when you have two media columns which are adding to a
single media field. For example, "File attachments (Filename)" and
"File attachments (URL)". If the columns appear in this order and you
only put values in the "File attachments (Filename)" column then when
the "File attachments (URL)" column is read after correctly reading
the files from the first column, the second column's values overwrite
it. If the second column is blank then you end up with nothing in your
field.