Decode facet values for display
In solr, the facet values seem to be stored using html encoding. This means
that when we show them in a facet sidebar (like in the gallery) we show the
&#xxx; value rather than the actual character. For example, entering an image
creator as "First O'Last" results in "First O&#039;Last" shown in the facet
UI. To resolve this, we just need to decode it in a couple of places.