Correctly look for existing backups
Backups are created using the site name from the sites/ path, however,
we look for existing backups to show the user using the domain of the
current request. These two things could be different resulting in the
admin page showing no backups even though there are some. Using the
same method that the backup is created to look for the backups resolves
this problem (we leave the old method there just in case we can't get
the public path).